refactor: User와 Auth 도메인으로 분리
